bulk-embeddings-sample/requirements.txt
asttokens==2.4.1
certifi==2024.2.2
cffi==1.16.0
circuitbreaker==1.4.0
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
greenlet==3.0.3
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
oci==2.126.0
oracledb==2.1.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.2
six==1.16.0
SQLAlchemy==2.0.29
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
wcwidth==0.2.13
